Sunglasses demo works (todo: rotate glasses)

+++ samples/funnyHats/js/ui.js
@@ -12,6 +12,8 @@ let glassesMask = null;
 let currentGlasses = 0;
 let glassesNum = document.getElementsByClassName('glasses-image').length;
 
+let objectChanged = false;
+
 
 function initUI() {
   initStats();
@@ -31,6 +33,7 @@ function initUI() {
       currentHat = i;
       hatSrc = hats[currentHat].src.clone();
       hatMask = hats[currentHat].mask.clone();
+      objectChanged = true;
     });
     // create mask from alpha channel
     cv.split(hatSrc, rgbaVector);
@@ -56,6 +59,7 @@ function initUI() {
       currentGlasses = i;
       glassesSrc = glasses[currentGlasses].src.clone();
       glassesMask = glasses[currentGlasses].mask.clone();
+      objectChanged = true;
     });
     // create mask from alpha channel
     cv.split(glassesSrc, rgbaVector);
@@ -96,14 +100,18 @@ function createNode(name, carouselName) {
 
 function deleteHats() {
   for (let i = 0; i < hats.length; i++) {
-    hats[i].src.delete();
-    hats[i].mask.delete();
+    if (hats[i].src != null && !hats[i].src.isDeleted())
+      hats[i].src.delete();
+    if (hats[i].mask != null && !hats[i].mask.isDeleted())
+      hats[i].mask.delete();
   }
 }
 
 function deleteGlasses() {
   for (let i = 0; i < glasses.length; i++) {
+    if (glasses[i].src != null && !glasses[i].src.isDeleted())
     glasses[i].src.delete();
+    if (glasses[i].mask != null && !glasses[i].mask.isDeleted())
     glasses[i].mask.delete();
   }
 }
